Power limit reject tests implemented
